+++ src/index.js
@@ -1,10 +1,12 @@
 import Container from './Container';
 import ServerList from './ServerListing';
 import { loadServers, announceEvents } from './hub';
+import useServers from './useServers';
 
 export {
   Container,
   ServerList,
   loadServers,
   announceEvents,
+  useServers,
 };