+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Use CSS instead of mui's styles mechanism.
 * Use material-ui 5.x alpha.
 * Load react-markdown once a description dialog is first opened.
+* Add a hook for typical server list state management.
 
 ## 4.0.0
 * Promote alpha to a full release.